refactor(ruby): extract solana-pay-core layer (PayCore namespace)
Mirrors the Rust spine umbrella layout (solana-pay-core / solana-mpp /
solana-x402 / solana-pay-kit). All shared Solana primitives + JCS RFC
8785 + RFC 7235 auth-param parser + RFC 3339 + base64url + canonical
L6 error codes live under PayCore::*. Both solana-mpp (Mpp::*) and
solana-x402 (X402::*) consume PayCore directly; no cross-layer
references.

  +------------------------------------------------------------+
  |                       solana-pay-kit                       |
  +---------------------------+--------------------------------+
  |        solana-mpp         |          solana-x402           |
  +---------------------------+--------------------------------+
  |                       solana-pay-core                      |
  +------------------------------------------------------------+

New files under ruby/lib/pay_core/:

- base64_url.rb        PayCore::Base64Url
- json.rb              PayCore::Json (RFC 8785 JCS)
- headers.rb           PayCore::Headers (generic RFC 7235 auth-param parser)
- rfc3339_parser.rb    PayCore::Rfc3339Parser
- error_codes.rb       PayCore::ErrorCodes (canonical L6 codes + classifier)
- solana/base58.rb     PayCore::Solana::Base58
- solana/programs.rb   PayCore::Solana::Programs (NEW: SYSTEM / TOKEN / TOKEN_2022 /
                       ASSOCIATED_TOKEN / MEMO / COMPUTE_BUDGET / LIGHTHOUSE)
- solana/caip2.rb      PayCore::Solana::Caip2 (NEW: MAINNET / DEVNET / TESTNET)
- solana/mints.rb      PayCore::Solana::Mints
- solana/public_key.rb PayCore::Solana::PublicKey (PDA derivation + on-curve)
- solana/ata.rb        PayCore::Solana::ATA (NEW name; was AssociatedToken)
- solana/account.rb    PayCore::Solana::Account
- solana/transaction.rb PayCore::Solana::Transaction (wire codec + short_vec helpers)
- solana/rpc.rb        PayCore::Solana::Rpc

Backward-compat alias layer (no public-API churn for MPP consumers):

- Mpp::Methods::Solana::Base58           = PayCore::Solana::Base58
- Mpp::Methods::Solana::Mints            = PayCore::Solana::Mints
- Mpp::Methods::Solana::PublicKey        = PayCore::Solana::PublicKey
- Mpp::Methods::Solana::Account          = PayCore::Solana::Account
- Mpp::Methods::Solana::AssociatedToken  = PayCore::Solana::ATA
- Mpp::Methods::Solana::Rpc              < PayCore::Solana::Rpc (overrides
                                           error class to raise Mpp::Error)
- Mpp::Methods::Solana::Transaction      < PayCore::Solana::Transaction
                                           (overrides sign_with error class to
                                            raise Mpp::VerificationError)
- Mpp::Methods::Solana::{Message, Instruction, AddressLookup, Cursor}
                                         = PayCore::Solana::*
- Mpp::Core::Base64Url                   = PayCore::Base64Url
- Mpp::Core::Json                        = PayCore::Json
- Mpp::Core::Rfc3339Parser               = PayCore::Rfc3339Parser
- Mpp::Core::Headers                     delegates to PayCore::Headers for
                                         generic auth-param parsing; keeps
                                         MPP-specific parse_www_authenticate /
                                         format_receipt / parse_receipt because
                                         they construct Mpp::Core::Challenge /
                                         Mpp::Core::Receipt
- Mpp::ErrorCodes                        = PayCore::ErrorCodes

solana-x402 (X402::Interop::*) now consumes PayCore directly; no
Mpp:: references remain in ruby/lib/x402/*.rb.

Umbrella ruby/lib/pay_kit.rb re-exports PayCore + Mpp + X402 under
PayKit::Core / PayKit::Mpp / PayKit::X402. Existing
`require "mpp"` and direct `require "x402/..."` paths still work.

Test results:

- bundle exec rake test: 229 runs, 770 assertions, 0 failures, 0 errors.
  214 baseline MPP tests preserved unchanged via aliases;
  15 new PayCore tests assert (a) PayCore::* are the canonical homes,
  (b) Mpp::* aliases resolve via assert_same,
  (c) Mpp::Methods::Solana::Rpc / Transaction subclass PayCore variants,
  (d) X402::Interop::Server::DEFAULT_NETWORK reads from
      PayCore::Solana::Caip2::DEVNET (no string literal duplicate).
- bundle exec standardrb: clean.

Test for `latest_blockhash_rejects_http_failure` updated to expect
PayCore::Solana::Rpc::RpcError (was previously expected to be
Mpp::Error, which only fires through the MPP charge path via the
subclass override).
