+++ ruby/lib/mpp/core/headers.rb
@@ -1,13 +1,19 @@
 # frozen_string_literal: true
 
+require "pay_core/headers"
+
 module Mpp
   module Core
-    # Parser and formatter for MPP HTTP headers.
+    # MPP-flavoured `Payment` header parser. Delegates the generic
+    # RFC 7235 auth-scheme/auth-param tokenisation to
+    # `PayCore::Headers`; only the MPP-specific bits (constructing a
+    # `Challenge` / `Receipt` from parsed params, choosing the canonical
+    # `Payment` scheme header name set) live in this module.
     module Headers
       WWW_AUTHENTICATE = "www-authenticate"
       AUTHORIZATION = "authorization"
       PAYMENT_RECEIPT = "payment-receipt"
-      PAYMENT_SCHEME = "Payment"
+      PAYMENT_SCHEME = ::PayCore::Headers::PAYMENT_SCHEME
 
       module_function
 
@@ -26,9 +32,11 @@ module Mpp
         "Payment #{parts.join(", ")}"
       end
 
-      # Parse all `Payment` challenges across one or more `WWW-Authenticate` values (RFC 7235 sec 4.1).
-      # Returns an array of successfully-parsed Challenge objects; malformed individual challenges are skipped.
-      # Mirrors the Rust spine which exposes Vec<Result<PaymentChallenge, Error>> and filters at the call site.
+      # Parse all `Payment` challenges across one or more
+      # `WWW-Authenticate` values (RFC 7235 sec 4.1). Returns an array of
+      # successfully-parsed Challenge objects; malformed individual
+      # challenges are skipped. Mirrors the Rust spine which exposes
+      # Vec<Result<PaymentChallenge, Error>> and filters at the call site.
       def parse_www_authenticate_all(headers)
         Array(headers).flat_map { |header| split_payment_challenge_values(header) }.filter_map do |chunk|
           parse_www_authenticate(chunk)
@@ -37,110 +45,20 @@ module Mpp
         end
       end
 
-      # Split a WWW-Authenticate header value into individual Payment challenges (quote-aware).
-      #
-      # Detects RFC 7235 sec 2.1 auth-scheme boundaries (a token followed by whitespace and a
-      # key=value pair), not just literal "Payment" occurrences. This is required to correctly
-      # terminate a Payment chunk when a different scheme (e.g. Bearer) follows it on the same
-      # header value, and to skip over non-Payment schemes that precede or interleave with
-      # Payment schemes.
+      # Generic auth-scheme splitter; delegates to PayCore.
       def split_payment_challenge_values(header)
-        bytes = header.to_s
-        scheme_starts = [] # array of [offset, is_payment]
-        in_quote = false
-        escaped = false
-        at_boundary = true
-        i = 0
-        while i < bytes.length
-          ch = bytes[i]
-          if in_quote
-            if escaped
-              escaped = false
-            elsif ch == "\\"
-              escaped = true
-            elsif ch == "\""
-              in_quote = false
-            end
-            i += 1
-            next
-          end
-
-          if ch == "\""
-            in_quote = true
-            at_boundary = false
-            i += 1
-            next
-          end
-
-          if ch == ","
-            at_boundary = true
-            i += 1
-            next
-          end
-
-          if [" ", "\t"].include?(ch)
-            i += 1
-            next
-          end
-
-          if at_boundary && token_char?(ch)
-            match = match_auth_scheme_start(bytes, i)
-            if match
-              scheme_end, is_payment = match
-              scheme_starts << [i, is_payment]
-              i = scheme_end
-              at_boundary = false
-              next
-            end
-          end
-
-          at_boundary = false
-          i += 1
-        end
-
-        return [] if scheme_starts.empty?
-
-        chunks = []
-        scheme_starts.each_with_index do |(start, is_payment), idx|
-          next unless is_payment
-
-          finish = scheme_starts[idx + 1] ? scheme_starts[idx + 1][0] : bytes.length
-          chunk = bytes[start...finish].strip.sub(/,\s*\z/, "").strip
-          chunks << chunk unless chunk.empty?
-        end
-        chunks
+        ::PayCore::Headers.split_payment_challenge_values(header)
       end
 
-      # RFC 7230 sec 3.2.6 tchar.
-      TCHAR_EXTRA = "!#$%&'*+-.^_`|~"
       def token_char?(ch)
-        return false unless ch
-
-        ch.match?(/[A-Za-z0-9]/) || TCHAR_EXTRA.include?(ch)
+        ::PayCore::Headers.token_char?(ch)
       end
 
-      # If `bytes[index]` starts an auth-scheme (RFC 7235 sec 2.1), return
-      # [offset_after_scheme, is_payment_scheme]. Otherwise return nil.
-      #
-      # A scheme requires: token, 1*SP, then non-empty content (either an
-      # auth-param list `key=val,...` or a token68 credential). A bare
-      # `token=` (no SP gap) is an auth-param continuation, not a new scheme.
       def match_auth_scheme_start(bytes, index)
-        token_end = index
-        token_end += 1 while token_end < bytes.length && token_char?(bytes[token_end])
-        return nil if token_end == index
-
-        return nil unless [" ", "\t"].include?(bytes[token_end])
-
-        cursor = token_end
-        cursor += 1 while cursor < bytes.length && [" ", "\t"].include?(bytes[cursor])
-        return nil if cursor >= bytes.length || bytes[cursor] == ","
-
-        scheme = bytes[index, token_end - index]
-        [token_end, scheme.casecmp(PAYMENT_SCHEME).zero?]
+        ::PayCore::Headers.match_auth_scheme_start(bytes, index)
       end
 
-      # Parse a single `WWW-Authenticate` challenge.
+      # Parse a single `WWW-Authenticate` challenge into a Challenge object.
       def parse_www_authenticate(header)
         params = parse_auth_params(strip_payment(header))
         request = params.fetch("request")
@@ -175,70 +93,19 @@ module Mpp
         )
       end
 
+      # Strip the leading "Payment " scheme tag from a header value.
       def strip_payment(header)
-        value = header.to_s.strip
-        scheme_len = PAYMENT_SCHEME.length
-        unless value.length > scheme_len && value[0, scheme_len].casecmp(PAYMENT_SCHEME).zero? && [" ", "\t"].include?(value[scheme_len])
-          raise ArgumentError, "expected Payment scheme"
-        end
-
-        value[(scheme_len + 1)..].strip
+        ::PayCore::Headers.strip_payment(header)
       end
 
-      # Parse RFC 7235 sec 2.1 auth-params; accepts quoted-string and token form.
+      # Parse RFC 7235 sec 2.1 auth-params; accepts quoted-string and
+      # token form. Delegates to PayCore::Headers.
       def parse_auth_params(input)
-        params = {}
-        index = 0
-        while index < input.length
-          index += 1 while index < input.length && [",", " ", "\t"].include?(input[index])
-          break if index >= input.length
-
-          key_start = index
-          index += 1 while index < input.length && input[index] != "=" && input[index] != "," && input[index] != " " && input[index] != "\t"
-          key = input[key_start...index]
-          index += 1 while index < input.length && [" ", "\t"].include?(input[index])
-          raise ArgumentError, "invalid auth parameter" if key.empty? || index >= input.length || input[index] != "="
-
-          index += 1
-          index += 1 while index < input.length && [" ", "\t"].include?(input[index])
-
-          value = if index < input.length && input[index] == "\""
-            index += 1
-            buf = +""
-            while index < input.length
-              char = input[index]
-              if char == "\\"
-                index += 1
-                buf << input[index].to_s
-              elsif char == "\""
-                index += 1
-                break
-              else
-                buf << char
-              end
-              index += 1
-            end
-            buf
-          else
-            value_start = index
-            index += 1 while index < input.length && input[index] != ","
-            input[value_start...index].rstrip
-          end
-
-          raise ArgumentError, "duplicate parameter: #{key}" if params.key?(key)
-          params[key] = value
-        end
-        params
+        ::PayCore::Headers.parse_auth_params(input)
       end
 
       def escape(value)
-        # RFC 9110 section 5.5 forbids CR and LF in header field values.
-        # Silent strip would let malformed inputs round-trip and would let a
-        # caller-controlled realm inject extra HTTP headers. Reject with an
-        # explicit error so the problem surfaces at emission time.
-        string = value.to_s
-        raise ArgumentError, "control character in header parameter value" if string.match?(/[\r\n]/)
-        string.gsub("\\", "\\\\\\").gsub("\"", "\\\"")
+        ::PayCore::Headers.escape(value)
       end
     end
   end
