+++ ruby/lib/mpp/methods/solana/associated_token.rb
@@ -1,24 +1,15 @@
 # frozen_string_literal: true
 
+require "pay_core/solana/ata"
+
 module Mpp
   module Methods
     module Solana
-      # Associated token account derivation helper.
-      module AssociatedToken
-        module_function
-
-        # Derive the ATA for owner/mint/token-program.
-        def derive(owner:, mint:, token_program:)
-          Solana::PublicKey.find_program_address(
-            [
-              Solana::PublicKey.new(owner).bytes.pack("C*"),
-              Solana::PublicKey.new(token_program).bytes.pack("C*"),
-              Solana::PublicKey.new(mint).bytes.pack("C*")
-            ],
-            Mints::ASSOCIATED_TOKEN_PROGRAM
-          ).first.to_s
-        end
-      end
+      # Backward-compat alias. Canonical home: `PayCore::Solana::ATA`.
+      # The class name stays `AssociatedToken` here because pre-PayCore
+      # MPP code imported it under that name; the underlying module is
+      # the same.
+      AssociatedToken = ::PayCore::Solana::ATA
     end
   end
 end
